backend/requirements.txt
autopep8==1.4.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
defusedxml==0.7.1
Django==2.2.13
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
oauthlib==3.1.1
pep8==1.7.1
pycodestyle==2.5.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2018.5
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.7
zipp==3.6.0
